Fixing existing tests
